+++ Minigin/Components/FPSComp.h
@@ -29,12 +29,12 @@ namespace enf
 		bool m_NeedsUpdate;
 		int m_Precision{ 1 };
 		float m_UpdateDelay{ .5f };
-		double m_Timer{};
+		float m_Timer{};
 		TextRenderComp* m_TextRenderCompPtr{};
 		std::shared_ptr<Font> m_FontPtr;
 
-		std::deque<double> m_TotalFPS{};
+		std::deque<float> m_TotalFPS{};
 
-		double GetAverageFps();
+		float GetAverageFps();
 	};
 }
